Move connection, result and relation to separate files
