+++ tools/pythonpkg/src/include/duckdb_python/pybind_wrapper.hpp
@@ -1,0 +1,18 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb_python/pybind_wrapper.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include <pybind11/pybind11.h>
+#include <pybind11/numpy.h>
+
+namespace duckdb {
+
+namespace py = pybind11;
+
+}
